+++ src/App.js
@@ -1,7 +1,8 @@
 import { Route } from 'react-router-dom'
 import { Login } from './component/login/Login'
-import { SignUp } from './component/singup/SignUp'
+import { SignUp } from './component/signup/SignUp'
 import { Home } from './component/home/Home';
+import './app.css'
 
 export function App() {
   return (
