+++ src/component/home/home.module.css
@@ -1,11 +1,3 @@
-.main{
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100vw;
-  height: 100vh;
-}
-
 .container{
   display: flex;
   flex-direction: column;
