+++ src/component/login/Login.js
@@ -1,30 +1,26 @@
 import React, { useState } from 'react';
 
-
 export function Login() {
-    const [input, setInput] = useState({ usuario: "", contrasena: "" })
+    const [user, setUser] = useState("")
+    const [password, setPassword] = useState("")
 
     const handleOnSubmit = (e) => {
         e.preventDefault()
-        console.log('usuario:' + input.usuario);
-        console.log('contrasena:' + input.contrasena);
-        setInput(input.usuario = "")
-        setInput(input.contrasena = "");
+
     }
 
     return (
-        <>
-            <form onSubmit={handleOnSubmit}>
-                <div>
-                    <input value={input.usuario} onChange={(e) => setInput(e.target.value)} />
-                    <label> Usuario or correo</label>
-                </div>
-                <div>
-                    <input value={input.contrasena} onChange={(e) => setInput(e.target.value)} />
-                    <label> Contrasena</label>
-                </div>
-                <input type="submit" value="ingresar" />
-            </form>
-        </>
+        <form onSubmit={handleOnSubmit}>
+            <h1>Login</h1>
+            <div>
+                <input value={user} onChange={(e) => setUser(e.target.value)} />
+                <label>Usuario</label>
+            </div>
+            <div>
+                <input value={password} onChange={(e) => setPassword(e.target.value)} />
+                <label> Contrasena</label>
+            </div>
+            <button>Ingresar</button>
+        </form>
     )
 }
